Encapsulate `IOEvent` and `DeviceMetadata`
